refactor: moving types
